add node tree method
